requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

channels==4.0.0
Django==4.2
djangorestframework==3.14.0
opencv-python==4.8.0.76
requests==2.31.0
